+++ vllm/entrypoints/pooling/classify/io_processor.py
@@ -1,30 +1,17 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 
-from collections.abc import Callable
-from typing import Any
 
-from vllm.entrypoints.chat_utils import (
-    ChatCompletionMessageParam,
-    ChatTemplateContentFormatOption,
-    ConversationMessage,
-)
-from vllm.entrypoints.openai.engine.serving import RendererChatRequest, RendererRequest
 from vllm.entrypoints.pooling.base.io_processor import PoolingIOProcessor
 from vllm.entrypoints.pooling.classify.protocol import (
     ClassificationChatRequest,
     ClassificationCompletionRequest,
 )
-from vllm.inputs import SingletonPrompt
-from vllm.renderers import merge_kwargs
 from vllm.renderers.inputs import TokPrompt
-from vllm.renderers.inputs.preprocess import parse_model_prompt, prompt_to_seq
-from vllm.tokenizers import TokenizerLike
-from vllm.tool_parsers import ToolParser
 
 
 class ClassifyIOProcessor(PoolingIOProcessor):
-    async def pre_process(
+    def pre_process(
         self, request: ClassificationCompletionRequest | ClassificationChatRequest
     ) -> list[TokPrompt] | None:
         if isinstance(request, ClassificationChatRequest):
@@ -33,7 +20,7 @@ class ClassifyIOProcessor(PoolingIOProcessor):
                 chat_template_kwargs=request.chat_template_kwargs,
                 trust_request_chat_template=self.trust_request_chat_template,
             )
-            _, engine_prompts = await self._preprocess_chat(
+            _, engine_prompts = self._preprocess_chat(
                 request,
                 request.messages,
                 default_template=self.chat_template,
@@ -41,7 +28,7 @@ class ClassifyIOProcessor(PoolingIOProcessor):
                 default_template_kwargs=None,
             )
         elif isinstance(request, ClassificationCompletionRequest):
-            engine_prompts = await self._preprocess_completion(
+            engine_prompts = self._preprocess_completion(
                 request,
                 prompt_input=request.input,
                 prompt_embeds=None,
@@ -49,101 +36,3 @@ class ClassifyIOProcessor(PoolingIOProcessor):
         else:
             raise ValueError("Invalid classification request type")
         return engine_prompts
-
-    async def _preprocess_completion(
-        self,
-        request: RendererRequest,
-        prompt_input: str | list[str] | list[int] | list[list[int]] | None,
-        prompt_embeds: bytes | list[bytes] | None,
-    ) -> list[TokPrompt]:
-        renderer = self.renderer
-        model_config = self.model_config
-
-        prompts = list[SingletonPrompt | bytes]()
-        if prompt_embeds is not None:  # embeds take higher priority
-            prompts.extend(prompt_to_seq(prompt_embeds))
-        if prompt_input is not None:
-            prompts.extend(prompt_to_seq(prompt_input))
-
-        parsed_prompts = [
-            (
-                prompt
-                if isinstance(prompt, bytes)
-                else parse_model_prompt(model_config, prompt)
-            )
-            for prompt in prompts
-        ]
-        tok_params = request.build_tok_params(model_config)
-
-        return await renderer.render_cmpl_async(
-            parsed_prompts,
-            tok_params,
-            prompt_extras={
-                k: v
-                for k in ("mm_processor_kwargs", "cache_salt")
-                if (v := getattr(request, k, None)) is not None
-            },
-        )
-
-    async def _preprocess_chat(
-        self,
-        request: RendererChatRequest,
-        messages: list[ChatCompletionMessageParam],
-        default_template: str | None,
-        default_template_content_format: ChatTemplateContentFormatOption,
-        default_template_kwargs: dict[str, Any] | None,
-        tool_dicts: list[dict[str, Any]] | None = None,
-        tool_parser: Callable[[TokenizerLike], ToolParser] | None = None,
-    ) -> tuple[list[ConversationMessage], list[TokPrompt]]:
-        from vllm.tokenizers.mistral import MistralTokenizer
-
-        renderer = self.renderer
-
-        default_template_kwargs = merge_kwargs(
-            default_template_kwargs,
-            dict(
-                tools=tool_dicts,
-                tokenize=isinstance(renderer.tokenizer, MistralTokenizer),
-            ),
-        )
-
-        tok_params = request.build_tok_params(self.model_config)
-        chat_params = request.build_chat_params(
-            default_template, default_template_content_format
-        ).with_defaults(default_template_kwargs)
-
-        (conversation,), (engine_prompt,) = await renderer.render_chat_async(
-            [messages],
-            chat_params,
-            tok_params,
-            prompt_extras={
-                k: v
-                for k in ("mm_processor_kwargs", "cache_salt")
-                if (v := getattr(request, k, None)) is not None
-            },
-        )
-
-        return conversation, [engine_prompt]
-
-    def _validate_chat_template(
-        self,
-        request_chat_template: str | None,
-        chat_template_kwargs: dict[str, Any] | None,
-        trust_request_chat_template: bool,
-    ):
-        if not trust_request_chat_template and (
-            request_chat_template is not None
-            or (
-                chat_template_kwargs
-                and chat_template_kwargs.get("chat_template") is not None
-            )
-        ):
-            raise ValueError(
-                "Chat template is passed with request, but "
-                "--trust-request-chat-template is not set. "
-                "Refused request with untrusted chat template."
-            )
-        return None
-
-    async def post_process(self):
-        pass
